Add session resume for combined healthcheck vacreg ussd
