+++ vaccine/tests/test_healthcheck_vacreg_ussd.py
@@ -2,7 +2,7 @@ import pytest
 
 from vaccine.healthcheck_ussd import Application as HealthCheckApp
 from vaccine.healthcheck_vacreg_ussd import Application
-from vaccine.models import Message, User
+from vaccine.models import Message, StateData, User
 from vaccine.vaccine_reg_ussd import Application as VacRegApp
 
 
@@ -27,3 +27,44 @@ async def test_menu():
     [reply] = await app.process_message(msg)
     assert len(reply.content) < 160
     assert u.state.name == "state_menu"
+
+
+@pytest.mark.asyncio
+async def test_session_timeout_healthcheck():
+    u = User(
+        addr="27820001001", state=StateData(name="state_sore_throat"), session_id=1
+    )
+    app = Application(u)
+    msg = Message(
+        content=None,
+        to_addr="27820001002",
+        from_addr="27820001001",
+        transport_name="whatsapp",
+        transport_type=Message.TRANSPORT_TYPE.HTTP_API,
+        session_event=Message.SESSION_EVENT.NEW,
+    )
+    [reply] = await app.process_message(msg)
+    assert u.state.name == "state_timed_out_healthcheck"
+    assert len(reply.content) < 140
+    assert u.answers["resume_state"] == "state_sore_throat"
+    assert u.session_id != 1
+
+
+async def test_state_timed_out_vaccinereg():
+    u = User(
+        addr="27820001001", state=StateData(name="state_vaccination_time"), session_id=1
+    )
+    app = Application(u)
+    msg = Message(
+        content=None,
+        to_addr="27820001002",
+        from_addr="27820001001",
+        transport_name="whatsapp",
+        transport_type=Message.TRANSPORT_TYPE.HTTP_API,
+        session_event=Message.SESSION_EVENT.NEW,
+    )
+    [reply] = await app.process_message(msg)
+    assert u.state.name == "state_timed_out_vacreg"
+    assert len(reply.content) < 140
+    assert u.answers["resume_state"] == "state_vaccination_time"
+    assert u.session_id != 1
