+++ vaccine/tests/test_vaccine_reg_ussd.py
@@ -1310,81 +1310,3 @@ async def test_state_error(evds_mock):
         "termsAndConditionsAccepted": True,
         "refugeeNumber": "6001010001081",
     }
-
-
-async def test_state_timed_out():
-    u = User(
-        addr="27820001001", state=StateData(name="state_medical_aid"), session_id=1
-    )
-    app = Application(u)
-    msg = Message(
-        content=None,
-        to_addr="27820001002",
-        from_addr="27820001001",
-        transport_name="whatsapp",
-        transport_type=Message.TRANSPORT_TYPE.HTTP_API,
-        session_event=Message.SESSION_EVENT.NEW,
-    )
-    [reply] = await app.process_message(msg)
-    assert u.state.name == "state_timed_out"
-    assert len(reply.content) < 140
-    assert u.answers["resume_state"] == "state_medical_aid"
-    assert u.session_id != 1
-
-
-async def test_state_timed_out_restart():
-    u = User(
-        addr="27820001001",
-        state=StateData(name="state_timed_out"),
-        session_id=1,
-        answers={"resume_state": "state_medical_aid"},
-    )
-    app = Application(u)
-    msg = Message(
-        content="start over",
-        to_addr="27820001002",
-        from_addr="27820001001",
-        transport_name="whatsapp",
-        transport_type=Message.TRANSPORT_TYPE.HTTP_API,
-    )
-    await app.process_message(msg)
-    assert u.state.name == "state_age_gate"
-
-
-async def test_state_timed_out_resume():
-    u = User(
-        addr="27820001001",
-        state=StateData(name="state_timed_out"),
-        session_id=1,
-        answers={"resume_state": "state_medical_aid"},
-    )
-    app = Application(u)
-    msg = Message(
-        content="continue where i left off",
-        to_addr="27820001002",
-        from_addr="27820001001",
-        transport_name="whatsapp",
-        transport_type=Message.TRANSPORT_TYPE.HTTP_API,
-    )
-    await app.process_message(msg)
-    assert u.state.name == "state_medical_aid"
-
-
-async def test_state_timed_out_error():
-    u = User(
-        addr="27820001001",
-        state=StateData(name="state_timed_out"),
-        session_id=1,
-        answers={"resume_state": "state_medical_aid"},
-    )
-    app = Application(u)
-    msg = Message(
-        content="invalid",
-        to_addr="27820001002",
-        from_addr="27820001001",
-        transport_name="whatsapp",
-        transport_type=Message.TRANSPORT_TYPE.HTTP_API,
-    )
-    [reply] = await app.process_message(msg)
-    assert u.state.name == "state_timed_out"
-    assert len(reply.content) < 160
